+++ src/components/ContactCard/ContactCard.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import ContactIdentity from "./ContactIdentity";
 import ContactFields from "./ContactFields";
-import contactPropTypes from "../ContactPropTypes";
+import { fullContactPropTypes } from "../ContactPropTypes";
 import {
   getFieldListFrom,
   filterFieldList,
@@ -31,15 +31,7 @@ const ContactCard = ({ title, contact, renderHeader, renderBody }) => {
 };
 
 ContactCard.propTypes = {
-  contact: PropTypes.shape({
-    name: contactPropTypes.name,
-    phone: PropTypes.arrayOf(contactPropTypes.phone),
-    email: PropTypes.arrayOf(contactPropTypes.email),
-    address: PropTypes.arrayOf(contactPropTypes.address),
-    cozy: PropTypes.arrayOf(contactPropTypes.cozy),
-    birthday: contactPropTypes.birthday,
-    note: contactPropTypes.note
-  }).isRequired,
+  contact: fullContactPropTypes.isRequired,
   title: PropTypes.string.isRequired,
   renderHeader: PropTypes.func,
   renderBody: PropTypes.func
